// src/app/auth/interfaces/ImvCeduladoMeducacion.ts
export interface ImvCeduladoMeducacion {
    cedula: number;
    cedulaMadre: number;
    cedulaPadre: number;
    codCondicionCedulado: number;
    codDomicilio: number;
    codEstadoCivil: number;
    codInstruccion: number;
    codLugarMatrimonio: number;
    codLugarNacimiento: number;
    codNacionalidad: number;
    codProfesion: string;
    codSexo: number;
    fechaExpedicionCed: Date;
    fechaFallecimiento: Date;
    fechaMatrimonio: Date;
    fechaNacimiento: Date;
    nombreCalle: string;
    nombreConyuge: string;
    nombreMadre: string;
    nombrePadre: string;
    nombres: string;
    numeroCasa: string;
}